// src/pages/statistics/statistics.enum.ts
export enum StatisticsEnum {
    FETCH_ALL = 'all',
    FETCH_LAST_WEEK = 'last week',
    FETCH_TWO_WEEKS = 'two weeks',

    RANGE = 'range',
    CHART_TYPE = 'doughnut',

    LABEL_ALL = 'Show All',
    LABEL_LAST_WEEK = 'Last Week',
    LABEL_TWO_WEEKS = 'Last Two Weeks',

    MESSAGE_OFFLINE = 'You are currently offline, unable to retrieve results',
    MESSAGE_ONLINE = 'You are back online, would like to retrieve results?',

    PLATFORM_IOS = 'ios',
    PLATFORM_ANDROID = 'android',
    PLATFORM_WINDOWS = 'windows',

    ALERT_TYPE = 'radio',
    ALERT_ONLINE = 'Online',
    ALERT_YES = 'Yes',
    ALERT_NO = 'No',
    ALERT_CANCEL = 'Cancel',
    ALERT_OK = 'Ok',
    ALERT_TITLE = 'Date Range',
    ALERT_ACTION_CANCELLED = 'Action Cancelled',

    LOADER_CONTENT = 'Loading Statistics',

    DATA_TOTAL_GAMES = 'totalGames',
    DATA_STATISTIC = 'Total Games Played',
    DATA_GOAL = 'Number Of Games',
    DATA_GOALS_AND_GAMES_COUNT = 'goalsAndGameCount'

}